Remove unnecessary library load line
After running into this[1] issue, I saw a proposed solution of removing
the offending line entirely.  Looking at the code it seems like that
should be fine to do — the current code does this:

    genome_sequence<-get(library(GTF_annotation$genome,character.only = TRUE))
    library(GTF_annotation$genome,character.only = TRUE)
    genome_sequence<-get(GTF_annotation$genome)

This fails because `library(…)` returns a vector of attached libraries,
not just the single one loaded, which `get` can't deal with.  `git
blame` says this line has been there since 2019, and I'm not sure how
this has ever worked.  Maybe no one has ever gone through this branch of
the enclosing `if` statement, or maybe R's `library` function changed at
some point?  I'm not sure, and I'm still a newbie at R so I don't know
how to nail that down further.

Returning to the code, the two lines following that problematic line
just seem to do effectively the same thing, but in two steps instead of
one (indentation mine):

    genome_sequence<-get(library(GTF_annotation$genome,character.only = TRUE))

                         library(GTF_annotation$genome,character.only = TRUE)
    genome_sequence<-get(GTF_annotation$genome)

So I think it's safe to just remove this line entirely.

[1] https://github.com/ohlerlab/RiboseQC/issues/21#issuecomment-1247016533